defined InteractiveApplication class
It's an abstract base class meant to be extended by the class
representing game, which draws its state on screen 30 times a second.
Checking input, calling to update state, drawing, and breaking the
game loop is requested by the base class, so it's an example of
Inversion of Control design.