implemented a stack of game windows
Application is finally capable of transitioning from main menu to the
level screen, when New Game item in the main menu is selected.
I couldn't use the call stack for that, because the game runs in a loop.
DungeonCrawlerGame and MainMenu classes, previously extending
InteractiveApplication, now extend GameWindow.
New class, Game, delegates calls to draw, update, and receiveInput
methods to the window that is focused (foreground window).
Game class also manages a stack of these windows, allowing other windows
to both: open a new window on top of itself, and close itself.