refactoring in GameLevelWindow class
Changed the type of `world` field from a two dimensional array to its
own class, `World`.
Acces to world element is done through `getElementLevel` method, which
implements `NullObject` pattern. It returns LevelElement.EMPTY object,
when caller requests level element outside of the bounds of the level.
Defined an iterator, returned by `allPoints` method , that generates
all points of the rectangle passed to the method, to take advantage
of enhanced for-loop feature of Java.